refact: rename stats package
